New folder tools
Added the tools folder to hold adittional tools for inteligov.

First tool added is the beta version of minifier, it uses yui and closure compiler for minify all the js or css in a directory tree.

It have options to search all posible minified files or just the css or just js and clean all the minified files.